+++ src/js/shapes/optionShape.js
@@ -1,10 +1,8 @@
 import PropTypes from 'prop-types';
 
 export default PropTypes.shape({
-    label: PropTypes.string.isRequired,
-    value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
-
-    // Optional properties
     disabled: PropTypes.bool,
+    label: PropTypes.string,
     title: PropTypes.string,
+    value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
 });